improve the code

+++ routers/admin.py
@@ -18,7 +18,7 @@ from fastapi import APIRouter
 from config.connection import get_db
 from config.config import https_url_for
 from config.config import pwd_context, hash_password
-from config.config import verify_session
+from config.config import verify_session, create_session_token
 
 # router  = APIRouter(prefix='/users', tags=['Users'])
 router  = APIRouter()
@@ -29,9 +29,14 @@ templates.env.globals["https_url_for"] = https_url_for
 
 UPLOAD_DIR = "uploaded_files"  # Directory to save uploaded files
 os.makedirs(UPLOAD_DIR, exist_ok=True)
+UPLOAD_PROFILE ="profiles"
+os.makedirs(UPLOAD_PROFILE, exist_ok=True)
 
 
-
+def check_admin(request: Request, db: Session) -> models.User:
+    username = request.cookies.get("username")
+    if not username:
+        raise HTTPException(status_code=403, detail="Access forbidden: Missing credentials")
 
 
 @router.get("/")
@@ -48,8 +53,14 @@ def login_user(request: Request, email:str=Form(...), password:str=Form(...), db
     if not user or not pwd_context.verify(password, user.hashed_password):
         return templates.TemplateResponse("login.html", {"request": request, "body_class": "bg-primary", "error": "Invalid username or password"})
     
+    session_token = create_session_token(
+        username=email,
+        ip=request.client.host,
+        user_agent=request.headers.get("user-agent"),
+
+    )
     response = RedirectResponse("/admin", status_code=303)
-    response.set_cookie(key="username", value=user.username, httponly=True, )  # Set cookie for 1 hour
+    response.set_cookie(key="session_token", value=session_token, httponly=True, secure=True,samesite="strict")  # Set cookie for 1 hour
     return response
 
 
@@ -62,9 +73,9 @@ def logout():
 @router.get('/admin', response_class=HTMLResponse)
 def index(request:Request,db:Session = Depends(get_db), auth:str=Depends(verify_session)):
     try:
-        username = request.cookies.get("username", "Guest")
-        users = db.query(models.User).all()
-        user = db.query(models.User).filter(models.User.username == username).first()
+
+        # users = db.query(models.User).all()
+        user = db.query(models.User).filter(models.User.email == auth).first()
         if not user:
             raise HTTPException(status_code=401, detail="User not found")
         
@@ -76,7 +87,7 @@ def index(request:Request,db:Session = Depends(get_db), auth:str=Depends(verify_
                     agent.date_of_birth,agent.birth_place,agent.category,agent.telephone, agent.document_path.replace("\\", "/")] for agent in agents]
         }
         
-        return templates.TemplateResponse("index.html",{"request":request, "body_class": "sb-nav-fixed", "data":data, "username":username, "role":role})
+        return templates.TemplateResponse("index.html",{"request":request, "body_class": "sb-nav-fixed", "data":data, "username":user.username, "role":role})
     except SQLAlchemyError as e:
     
         raise HTTPException(status_code=500, detail="A database error occurred.")
@@ -84,42 +95,39 @@ def index(request:Request,db:Session = Depends(get_db), auth:str=Depends(verify_
     except Exception as e:
                 # Log unexpected errors
         print(f"Unexpected error: {str(e)}")
-        raise HTTPException(status_code=500, detail="An unexpected error occurred.")
+        raise HTTPException(status_code=500, detail="Quelque chose ne va pas! Ressayer encore!! .")
     
 
 
 @router.get("/admin/users-table", response_class=HTMLResponse)
-async def users_table(request: Request,db: Session = Depends(get_db),auth:str=Depends(verify_session)):
-    username = request.cookies.get("username", None)
+async def users_table(request: Request,db: Session = Depends(get_db),username:str=Depends(verify_session)):
+    username = request.cookies.get("username")
 
     # Fetch the user from the database based on the username
-    user = db.query(models.User).filter(models.User.username == username).first()
+    user = db.query(models.User).filter(models.User.email == username).first()
     
     # Check if the user exists and if their role is 'admin'
     if not user or user.role != 'admin':
-        raise HTTPException(status_code=403, detail="Access forbidden: Admins only.")
+        raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
     
     users = db.query(models.User).all()
     data = {
         "columns": ["ID", "First Name", "Last Name", "Username", "Email", "Gender", "Is active"],  # Adjust based on your User model
         "rows": [[user.id, user.first_name, user.last_name, user.username, user.email,user.gender,user.role] for user in users]
     }
+
     return templates.TemplateResponse("users_table.html", {"request": request, "body_class": "sb-nav-fixed", "data":data})
 
 
 
 
 @router.get("/admin/agents-table", response_class=HTMLResponse)
-async def agents_table(request: Request,db: Session = Depends(get_db),auth:str=Depends(verify_session)):
-    username = request.cookies.get("username", None)
+async def agents_table(request: Request,db: Session = Depends(get_db),username:str=Depends(verify_session)):
 
-    if not username:
-        raise HTTPException(status_code=403, detail="Access forbidden: Missing credentials.")
-
-    admin_user = db.query(models.User).filter(models.User.username == username).first()
+    user = db.query(models.User).filter(models.User.email == username).first()
     # Check if the user exists and if their role is 'admin'
-    if not admin_user or admin_user.role != 'admin':
-        raise HTTPException(status_code=403, detail="Access forbidden: Admins only.")
+    if not user or user.role != 'admin':
+        raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
     
     agents = db.query(models.Agent).all()
     table_data = {
@@ -134,12 +142,11 @@ async def agents_table(request: Request,db: Session = Depends(get_db),auth:str=D
                   ] 
                   for agent in agents]
     }
-    print(admin_user.role)
     return templates.TemplateResponse("agents_table.html", 
                                       {"request": request, 
                                        "body_class": "sb-nav-fixed", 
                                        "data":table_data,
-                                       "role":admin_user.role})
+                                       "role":user.role})
 
 
 
@@ -156,8 +163,17 @@ def create_user(request:Request,
                 gender:str = Form(...),
                 db:Session = Depends(get_db),
                 auth:str=Depends(verify_session)):
+    # usernames= request.cookies.get("username")
     
     try:
+         
+        # if not usernames:
+        #     raise HTTPException(status_code=403, detail="Access forbidden: Missing credentials.")
+
+        admin_user = db.query(models.User).filter(models.User.email== auth).first()
+    # Check if the user exists and if their role is 'admin'
+        if not admin_user or admin_user.role != 'admin':
+            raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
         hashed_pw = hash_password(password)
         new_user = models.User(
             first_name=first_name,
@@ -181,12 +197,12 @@ def create_user(request:Request,
         if "psycopg2.errors.UniqueViolation" in str(e):
             raise HTTPException(
                 status_code=StarletteHTTPException.status_code, 
-                detail="A record with this NNI or this Numero de titre already exists. Please verify the data and try again."
+                detail="Verifier votre email et ressayer"
             )
     
         raise HTTPException(
             status_code=StarletteHTTPException.status_code, 
-            detail="An unexpected error occurred. Please contact the administrator."
+            detail="Une erreur inattendue s'est produite. Veuillez contacter l'administrateur."
 
         )
 
@@ -209,18 +225,38 @@ async def create_agent(request:Request,
                  telephone:str=Form(...),
                  address:str=Form(...),
                  document: UploadFile = File(...),
+                 profile:UploadFile =File(None),
                  db:Session=Depends(get_db),
                  auth:str=Depends(verify_session)):
+    
+    # username = request.cookies.get("username")
+
+    # if not username:
+    #     raise HTTPException(status_code=403, detail="Access forbidden: Missing credentials.")
+    
+
+    admin_user = db.query(models.User).filter(models.User.email == auth).first()
+    # Check if the user exists and if their role is 'admin'
+    if not admin_user or admin_user.role != 'admin':
+        raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
+
     form_data = await request.form()
     error_message = None
     file_path = f"{UPLOAD_DIR}/{document.filename}"#Path(UPLOAD_DIR) /document.filename
     with open(file_path, 'wb') as file:
         content = await document.read()
         file.write(content)
+
+    if profile : 
+        profile_image_path =f"{UPLOAD_PROFILE}/{profile.filename}"
+        with open(profile_image_path, 'wb') as f:
+            content = await profile.read()
+            f.write(content)
+
+    else:
+        profile_image_path =f"{UPLOAD_PROFILE}/profile_default.png"
+    
     try:
-        # file_path = f"uploaded_file/{document.filename}"
-        # with open(file_path, "wb") as file:
-        #      file.write( await document.read())
         new_agent = models.Agent(
                 nni=nni, 
                 title_number=title_number, 
@@ -230,7 +266,8 @@ async def create_agent(request:Request,
                 birth_place=birth_place,
                 telephone=telephone,
                 address=address,
-                document_path = str(file_path)
+                document_path = str(file_path),
+                profile_path =str(profile_image_path) 
                 )
         db.add(new_agent)
         db.commit()
@@ -242,10 +279,10 @@ async def create_agent(request:Request,
     
     except IntegrityError as e:
             db.rollback()
-            if "ix_agents_nni" in str(e.orig):
-                error_message = "This NNI already exists. Please try a different one."
+            if "ix_agents_nni" in str(e.orig) or "ix_agents_title_number" in str(e.orig):
+                error_message = "Un enregistrement avec ce NNI ou ce Numéro de titre existe déjà. Veuillez vérifier les données et réessayer."
             else:
-                error_message = "An unexpected error occurred. Please try again later."
+                error_message = "Une erreur inattendue s'est produite. Veuillez réessayer plus tard."
  
     return templates.TemplateResponse(
             "create_agent.html",
@@ -255,8 +292,17 @@ async def create_agent(request:Request,
    
 
 @router.delete("/admin/delete-agent/{agent_id}")
-async def delete_agent(agent_id: int, db: Session = Depends(get_db),auth:str=Depends(verify_session)):
+async def delete_agent(request:Request,agent_id: int, db: Session = Depends(get_db),auth:str=Depends(verify_session)):
     agent = db.query(models.Agent).filter(models.Agent.id == agent_id).first()
+    username = request.cookies.get("username")
+
+    if not username:
+        raise HTTPException(status_code=403, detail="Access forbidden: Missing credentials.")
+
+    admin_user = db.query(models.User).filter(models.User.username == username).first()
+    # Check if the user exists and if their role is 'admin'
+    if not admin_user or admin_user.role != 'admin':
+        raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
     if not agent:
         raise HTTPException(status_code=404, detail="Agent not found.")
     db.delete(agent)
@@ -265,8 +311,17 @@ async def delete_agent(agent_id: int, db: Session = Depends(get_db),auth:str=Dep
 
 
 @router.delete("/admin/delete-user/{username}", status_code=status.HTTP_200_OK)
-def delete_user(username:int, db:Session = Depends(get_db),auth:str=Depends(verify_session)):
+def delete_user(request:Request, username:int, db:Session = Depends(get_db),auth:str=Depends(verify_session)):
     user = db.query(models.User).filter(models.User.username==username).first()
+    username = request.cookies.get("username")
+
+    if not username:
+        raise HTTPException(status_code=403, detail="Access forbidden: Missing credentials.")
+
+    admin_user = db.query(models.User).filter(models.User.username == username).first()
+    # Check if the user exists and if their role is 'admin'
+    if not admin_user or admin_user.role != 'admin':
+        raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
     db.delete(user)
@@ -280,7 +335,13 @@ def delete_user(username:int, db:Session = Depends(get_db),auth:str=Depends(veri
 # Edit Agent section
 
 @router.get("/admin/get-agent/{agent_id}")
-async def get_agent(agent_id: int, db: Session = Depends(get_db)):
+async def get_agent(request:Request,agent_id: int, db: Session = Depends(get_db), auth:str=Depends(verify_session)):
+    # Fetch the user from the database based on the username
+    user = db.query(models.User).filter(models.User.email == auth).first()
+    
+    # Check if the user exists and if their role is 'admin'
+    if not user or user.role != 'admin':
+        raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
     agent = db.query(models.Agent).filter(models.Agent.id == agent_id).first()
     if not agent:
         raise HTTPException(status_code=404, detail="Agent not found.")
@@ -294,20 +355,71 @@ async def get_agent(agent_id: int, db: Session = Depends(get_db)):
         "category": agent.category,
         "address": agent.address,
         "telephone": agent.telephone,
+        # "document": agent.document_path.replace("\\", "/")  # Replace backslashes with forward slashes for correct file path in browser.
+        
     }
+
+@router.get("/admin/detail-agent/{agent_id}", response_class=HTMLResponse)
+async def agent_details(request:Request,agent_id: int, db: Session = Depends(get_db), auth:str =Depends(verify_session)):
+    try:
+      
+        admin_user = db.query(models.User).filter(models.User.email == auth).first()
+        agent = db.query(models.Agent).filter(models.Agent.id == agent_id).first()
+        
+        if not admin_user or admin_user.role != 'admin':
+            raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
+
+        if not agent:
+            raise HTTPException(status_code=404, detail="Cet agent n'existe pas dans la base de données.")
+        agent_data =  {
+            "id": agent.id,
+            "nni": agent.nni,
+            "title_number": agent.title_number,
+            "fullname": agent.fullname,
+            "date_of_birth": agent.date_of_birth,
+            "birth_place": agent.birth_place,
+            "category": agent.category,
+            "address": agent.address,
+            "telephone": agent.telephone,
+            "document": agent.document_path.replace("\\", "/"),# Replace backslashes with forward slashes for correct file path in browser.
+            "profile":agent.profile_path#.replace("\\", "/"),# Replace backslashes with forward slashes
+            
+        }
+
+        return templates.TemplateResponse("agent_details.html", {"request": request, "agent": agent_data})
+    
+    except Exception as e:
+        raise HTTPException(
+            status_code=StarletteHTTPException.status_code, 
+            detail=str(e)
+        )
+
+
 
 
 @router.put("/admin/edit-agent/{agent_id}")
-async def edit_agent(agent_id: int, updated_data: dict, db: Session = Depends(get_db)):
-    agent = db.query(models.Agent).filter(models.Agent.id == agent_id).first()
-    if not agent:
-        raise HTTPException(status_code=404, detail="Agent not found.")
+async def edit_agent(request:Request,agent_id: int, updated_data: dict, db: Session = Depends(get_db), auth:str=Depends(verify_session)):
+    try:
+        
+        admin_user = db.query(models.User).filter(models.User.email == auth).first()
+        agent = db.query(models.Agent).filter(models.Agent.id == agent_id).first()
     
-    for key, value in updated_data.items():
-        setattr(agent, key, value)
-    db.commit()
-    db.refresh(agent)
-    return {"message": "Agent updated successfully."}
+    
+        if not admin_user or admin_user.role != 'admin':
+            raise HTTPException(status_code=403, detail="Accès interdit : Réservé aux administrateurs uniquement.")
+        if not agent:
+            raise HTTPException(status_code=404, detail="Agent inexistant.")
+        
+        for key, value in updated_data.items():
+            setattr(agent, key, value)
+        db.commit()
+        db.refresh(agent)
+        return {"message": "Mise à jour correcte!."}
+    except Exception as e:
+        raise HTTPException(
+            status_code=StarletteHTTPException.status_code, 
+            detail="Nous ne pouvons pas modifier cet agent. Contacter l'administrateur"
+        )
